misc: replaced Reducer with Catalog and separated photometry. Changed previous Photometer to DifferentialPhotometer for clarity. Removed option to specify type of photometry (e.g., 'normal' or 'optimal') and now the behaviour is equivalent to 'both'. Also added pixel scale information for cameras so PSFs are now logged in arcseconds as well as pixels.
